feat(projects): Allow multiple getters, and other new features.
Allows each "getter" in a project to be a single getter or an array
of getters.  This allows use to do things like check for `volta` when
trying to get the node verson, and falling back to `node --version`
if volta is not installed.  We also allow specifying `cache.files` in
a getter, to have a list of files we should check have not changed.

Allow styles in default project config.  These can be overridden by
the "project" module.